feat: added basic testing structure BREAKING CHANGE: Not complete
