+++ docker-compose.test.yml
@@ -8,45 +8,30 @@ services:
     ports:
       - 3001:3000
     restart: always
-    env_file: 
-      - .env
+    env_file:
+      - test.env
     volumes:
       - ./:/app
       - /app/node_modules
-      
-  test-classicmodels-db:
+    depends_on:
+      - db
+
+  db:
     image: mysql:8.0
     restart: always
-    env_file: 
-    - .env
+    env_file:
+      - test.env
     cap_add:
       - SYS_NICE
     environment:
       - MYSQL_DATABASE=classicmodels
       - MYSQL_ROOT_PASSWORD=${DB_PASSWORD}
     ports:
-      - '3308:3306'
+      - '3309:3306'
     volumes:
       - db:/var/lib/mysql
       - ./db/init.sql:/docker-entrypoint-initdb.d/init.sql
 
-  test-classicmodels-schema-db:
-    image: mysql:8.0
-    restart: always
-    env_file: 
-    - .env
-    cap_add:
-      - SYS_NICE
-    environment:
-      - MYSQL_DATABASE=classicmodels-schema
-      - MYSQL_ROOT_PASSWORD=${DB_PASSWORD}
-    ports:
-      - '3309:3306'
-    volumes:
-      - db-schema:/var/lib/mysql
-      - ./db/schema.sql:/docker-entrypoint-initdb.d/schema.sql
 volumes:
   db:
-    driver: local
-  db-schema:
     driver: local
