+++ docker-compose.yml
@@ -8,19 +8,19 @@ services:
     ports:
       - 3000:3000
     restart: always
-    env_file: 
+    env_file:
       - .env
     volumes:
       - ./:/app
       - /app/node_modules
     depends_on:
-    - db
-      
+      - db
+
   db:
     image: mysql:8.0
     restart: always
-    env_file: 
-    - .env
+    env_file:
+      - .env
     cap_add:
       - SYS_NICE
     environment:
