+++ package.json
@@ -25,6 +25,7 @@
     "@commitlint/config-conventional": "^17.4.4",
     "@types/express": "^4.17.17",
     "@types/node": "^18.15.11",
+    "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^5.57.0",
     "eslint": "^8.37.0",
     "eslint-config-prettier": "^8.8.0",
@@ -36,6 +37,7 @@
     "husky": "^8.0.0",
     "nodemon": "^2.0.22",
     "prisma": "^4.12.0",
+    "supertest": "^6.3.3",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.3"
   }
