+++ src/routes/product.ts
@@ -45,20 +45,17 @@ productRouter.post('/products', async (req, res) => {
     })
 
     res.json({ product })
-
   } catch (e) {
     if (e instanceof Prisma.PrismaClientKnownRequestError) {
       // The .code property can be accessed in a type-safe manner
       if (e.code === 'P2002') {
         return res.status(409).json({
-          error: 'There is already a product with this code'
+          error: 'There is already a product with this code',
         })
       }
     }
     throw e
   }
-  
-
 })
 
 productRouter.get(
